bump versions, hoist stuff to index for export
