+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@jupyrdf/jupyter-elk",
-  "version": "0.3.0",
+  "version": "1.0.0",
   "description": "ElkJS widget for Jupyter",
   "keywords": [
     "jupyter",
